move vendored libraries
