+++ src/database/sqlite/sqlite_db_migration.h
@@ -3,7 +3,7 @@
 #define NETDATA_SQLITE_DB_MIGRATION_H
 
 #include "daemon/common.h"
-#include "database/sqlite/sqlite3/sqlite3.h"
+#include "database/sqlite/vendored/sqlite3.h"
 
 int perform_database_migration(sqlite3 *database, int target_version);
 int perform_context_database_migration(sqlite3 *database, int target_version);
